+++ src/main/java/com/yapp/artie/domain/archive/service/ArtworkService.java
@@ -11,7 +11,7 @@ import com.yapp.artie.domain.archive.exception.ArtworkNotFoundException;
 import com.yapp.artie.domain.archive.repository.ArtworkRepository;
 import com.yapp.artie.domain.s3.service.S3Service;
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.domain.user.deprecated.LoadUserJpaEntityApi;
+import com.yapp.artie.global.deprecated.LoadUserJpaEntityApi;
 import com.yapp.artie.global.util.S3Utils;
 import java.util.List;
 import java.util.stream.Collectors;
