+++ src/main/java/com/yapp/artie/domain/archive/service/CategoryService.java
@@ -12,7 +12,7 @@ import com.yapp.artie.domain.archive.exception.NotOwnerOfCategoryException;
 import com.yapp.artie.domain.archive.repository.CategoryRepository;
 import com.yapp.artie.domain.archive.repository.ExhibitRepository;
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.domain.user.deprecated.LoadUserJpaEntityApi;
+import com.yapp.artie.global.deprecated.LoadUserJpaEntityApi;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
